OpenAi Chatbot/requirements.txt
langchain_openai
langchain
python-dotenv
langchain_community
python-dotenv
langchain_core
streamlit
